Unify implementations of StringFormat.
